hpcgap: merge src/gvars.c
